+++ vowpalwabbit/io.cc
@@ -7,7 +7,7 @@ embodied in the content of this file are licensed under the BSD
 
 #include "io.h"
 
-unsigned int buf_read(io_buf &i, char* &pointer, int n)
+size_t buf_read(io_buf &i, char* &pointer, size_t n)
 {//return a pointer to the next n bytes.  n must be smaller than the maximum size.
   if (i.space.end + n <= i.endloaded)
     {
@@ -19,7 +19,7 @@ unsigned int buf_read(io_buf &i, char* &pointer, int n)
     {
       if (i.space.end != i.space.begin) //There exists room to shift.
 	{ // Out of buffer so swap to beginning.
-	  int left = i.endloaded - i.space.end;
+	  size_t left = i.endloaded - i.space.end;
 	  memmove(i.space.begin, i.space.end, left);
 	  i.space.end = i.space.begin;
 	  i.endloaded = i.space.begin+left;
