+++ vowpalwabbit/unique_sort.cc
@@ -8,7 +8,7 @@ int order_features(const void* first, const void* second)
 
 int order_audit_features(const void* first, const void* second)
 {
-  return ((audit_data*)first)->weight_index - ((audit_data*)second)->weight_index;
+  return (int)(((audit_data*)first)->weight_index) - (int)(((audit_data*)second)->weight_index);
 }
 
 void unique_features(v_array<feature> &features)
